@@ -8,45 +8,50 @@
     "babel-polyfill": "^6.26.0",
     "bluebird": "^3.5.1",
     "body-parser": "^1.18.2",
-    "office-ui-fabric-react": "5.25.0",
-    "react": "16.1.1",
-    "react-dom": "16.1.1",
+    "int-to-rgb": "^1.0.0",
+    "lumi-aqara": "^1.4.0",
+    "miio": "^0.14.0",
+    "mongodb": "^3.0.1",
+    "node-uuid": "^1.4.8",
+    "office-ui-fabric-react": "^5.35.2",
+    "react": "16.2.0",
+    "react-dom": "16.2.0",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "serve-favicon": "^2.4.5",
     "typescript": "^2.6.2",
-    "webpack": "^3.8.1"
+    "webpack": "^3.10.0"
   },
   "devDependencies": {
     "@types/axios": "^0.14.0",
-    "@types/bluebird": "^3.5.18",
+    "@types/bluebird": "^3.5.19",
     "@types/debug": "0.0.30",
-    "@types/express": "4.0.39",
+    "@types/express": "4.11.0",
     "@types/microsoft-ajax": "0.0.33",
-    "@types/node": "^8.0.53",
-    "@types/react": "16.0.25",
+    "@types/mongodb": "^2.2.18",
+    "@types/node": "^8.5.2",
+    "@types/react": "^16.0.31",
     "@types/react-dom": "16.0.3",
-    "@types/react-router": "^4.0.19",
+    "@types/react-router": "^4.0.20",
     "@types/react-router-dom": "^4.2.3",
     "@types/requirejs": "2.1.31",
-    "awesome-typescript-loader": "^3.4.0",
+    "awesome-typescript-loader": "^3.4.1",
     "axios": "^0.17.1",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
-    "compression-webpack-plugin": "^1.0.1",
-    "copy-webpack-plugin": "4.2.3",
+    "compression-webpack-plugin": "^1.1.3",
+    "copy-webpack-plugin": "4.3.1",
     "css-loader": "^0.28.7",
     "debug": "^3.1.0",
-    "ejs": "^2.5.7",
     "express": "^4.16.2",
-    "file-loader": "1.1.5",
+    "file-loader": "1.1.6",
     "json-loader": "^0.5.7",
     "source-map-loader": "^0.2.3",
-    "style-loader": "^0.19.0",
-    "ts-loader": "3.1.1",
+    "style-loader": "^0.19.1",
+    "ts-loader": "3.2.0",
     "tslint": "^5.8.0",
     "tslint-loader": "^3.5.3",
     "url-loader": "^0.6.2",
