@@ -3,10 +3,13 @@ import { Route, Switch, Link } from "react-router-dom";
 import { Application } from "./application";
 import { Application as YeelightApplication } from "../../../projects/yeelight/components/pages/application";
 import { Application as VacuumApplication } from "../../../projects/vacuumRoboter/components/pages/application";
-import { Application as AldiApplication } from "../../../projects/aldi/components/pages/application";
+import { Application as AldiApplication } from "../../../projects/aldi/components/pages/Application";
+import { Application as SensorApplication } from "../../../projects/xiaomi/components/pages/sensors";
+import { Application as GatewayApplication } from "../../../projects/xiaomi/components/pages/gateways";
 import { NotFoundPage } from "../../components/simple/NotFoundPage";
-import { RedirectWithStatus } from "../simple/routing";
+import { RedirectWithStatus } from "../simple/Routing";
 import { BasePage } from "../container/basePage";
+import { SystemInfo } from "../../../projects/system/components/pages/systeminfo";
 
 export interface IGlobalApplicationProps {
     requestUrl: string;
@@ -21,26 +24,35 @@ export class GlobalApplication extends React.Component<IGlobalApplicationProps, 
     }
     render() {
         console.log("render Application");
-        return <BasePage Body={
-            <div>
-                <h1>Hello from GlobalApplication!<br />Your requested url is {this.props.requestUrl}</h1>
-                <Switch>
-                    <RedirectWithStatus status={302} from="/courses" to="/aldi" />
-                    <Route exact path="/" component={Application} key="r1" />
-                    <Route path="/light" component={YeelightApplication} key="r2" />
-                    <Route path="/aldi" component={AldiApplication} key="r3" />
-                    <Route path="/vacuum" component={VacuumApplication} key="r4" />
-                    <Route path="*" component={NotFoundPage} key="r5" />
-                </Switch>
-            </div>
+        return <BasePage IncludeFabricElement={true} Body={
+            <Switch>
+                <RedirectWithStatus status={302} from="/courses" to="/aldi" />
+                <Route exact path="/" component={Application} key="r1" />
+                <Route path="/system" component={SystemInfo} key="r8" />
+                <Route path="/light" component={YeelightApplication} key="r2" />
+                <Route path="/aldi" component={AldiApplication} key="r3" />
+                <Route path="/vacuum" component={VacuumApplication} key="r4" />
+                <Route path="/sensors" component={SensorApplication} key="r6" />
+                <Route path="/gateways" component={GatewayApplication} key="r7" />
+                <Route path="*" component={NotFoundPage} key="r5" />
+            </Switch>
         } Navigation={
             <div>
                 <ul>
                     <li>
-                        <Link to={"/"} replace={true}>Homepage</Link>
+                        <Link to={"/"} replace={true}>Übersicht</Link>
+                    </li>
+                    <li>
+                        <Link to={"/system"} replace={true}>System-Informationen</Link>
                     </li>
                     <li>
                         <Link to={"/light"} replace={true}>Yeelight</Link>
+                    </li>
+                    <li>
+                        <Link to={"/sensors"} replace={true}>Sensoren</Link>
+                    </li>
+                    <li>
+                        <Link to={"/gateways"} replace={true}>Gateways</Link>
                     </li>
                     <li>
                         <Link to={"/aldi"} replace={true}>Aldi</Link>
